+++ src/constants/firestore.js
@@ -1,5 +1,4 @@
 export const SYNC_EVENTS = 'SYNC_EVENTS';
-export const SET_EVENTS = 'SET_EVENTS';
 export const CREATE_EVENT = 'CREATE_EVENT';
 export const UPDATE_EVENT = 'UPDATE_EVENT';
 export const DELETE_EVENT = 'DELETE_EVENT';
